requirements.txt
web3>=6.0
requests
tqdm
pandas
python-dotenv
